requirements.txt
ollama>=0.4.0
python-dotenv>=1.0.0
requests>=2.31.0
ddgs>=8.0.0
